Make the interpreters module a package.
